move : 폴더 이동 - 로그인 페이지 - 회원가입 페이지 간 이동을 위해 pages 폴더로 이동
